+++ compiler-cli/src/dependencies.rs
@@ -20,6 +20,7 @@ use gleam_core::{
 };
 use hexpm::version::Version;
 use itertools::Itertools;
+use same_file::is_same_file;
 use smol_str::SmolStr;
 use strum::IntoEnumIterator;
 
@@ -511,7 +512,7 @@ fn get_manifest<Telem: Telemetry>(
     };
 
     if should_resolve {
-        let manifest = resolve_versions(runtime, mode, config, None, telemetry)?;
+        let manifest = resolve_versions(runtime, mode, paths, config, None, telemetry)?;
         return Ok((true, manifest));
     }
 
@@ -524,7 +525,7 @@ fn get_manifest<Telem: Telemetry>(
         Ok((false, manifest))
     } else {
         tracing::debug!("manifest_outdated");
-        let manifest = resolve_versions(runtime, mode, config, Some(&manifest), telemetry)?;
+        let manifest = resolve_versions(runtime, mode, paths, config, Some(&manifest), telemetry)?;
         Ok((true, manifest))
     }
 }
@@ -536,7 +537,7 @@ struct ProvidedPackage {
     requirements: HashMap<SmolStr, hexpm::version::Range>,
 }
 
-#[derive(Clone, Eq, PartialEq, Debug)]
+#[derive(Clone, Eq, Debug)]
 enum ProvidedPackageSource {
     Git { repo: SmolStr, commit: SmolStr },
     Local { path: PathBuf },
@@ -612,9 +613,35 @@ impl ProvidedPackageSource {
     }
 }
 
+impl PartialEq for ProvidedPackageSource {
+    fn eq(&self, other: &ProvidedPackageSource) -> bool {
+        match (self, other) {
+            (
+                ProvidedPackageSource::Local { path: own_path },
+                ProvidedPackageSource::Local { path: other_path },
+            ) => is_same_file(own_path, other_path).unwrap_or(false),
+
+            (
+                ProvidedPackageSource::Git {
+                    repo: own_repo,
+                    commit: own_commit,
+                },
+                ProvidedPackageSource::Git {
+                    repo: other_repo,
+                    commit: other_commit,
+                },
+            ) => own_repo == other_repo && own_commit == other_commit,
+
+            (ProvidedPackageSource::Git { .. }, ProvidedPackageSource::Local { .. })
+            | (ProvidedPackageSource::Local { .. }, ProvidedPackageSource::Git { .. }) => false,
+        }
+    }
+}
+
 fn resolve_versions<Telem: Telemetry>(
     runtime: tokio::runtime::Handle,
     mode: Mode,
+    project_paths: &ProjectPaths,
     config: &PackageConfig,
     manifest: Option<&Manifest>,
     telemetry: &Telem,
@@ -635,12 +662,16 @@ fn resolve_versions<Telem: Telemetry>(
             Requirement::Path { path } => provide_local_package(
                 SmolStr::from(&name),
                 &path,
+                project_paths,
                 &mut provided_packages,
                 &mut vec![],
             )?,
-            Requirement::Git { git } => {
-                provide_git_package(SmolStr::from(&name), &git, &mut provided_packages)?
-            }
+            Requirement::Git { git } => provide_git_package(
+                SmolStr::from(&name),
+                &git,
+                project_paths,
+                &mut provided_packages,
+            )?,
         };
         let _ = root_requirements.insert(name, version);
     }
@@ -678,19 +709,18 @@ fn resolve_versions<Telem: Telemetry>(
 fn provide_local_package(
     package_name: SmolStr,
     package_path: &Path,
+    project_paths: &ProjectPaths,
     provided: &mut HashMap<SmolStr, ProvidedPackage>,
     parents: &mut Vec<SmolStr>,
 ) -> Result<hexpm::version::Range> {
-    let canonical_path = package_path
-        .canonicalize()
-        .map_err(|_| Error::DependencyCanonicalizationFailed(package_name.to_string()))?;
     let package_source = ProvidedPackageSource::Local {
-        path: canonical_path.clone(),
+        path: package_path.to_path_buf(),
     };
     provide_package(
         package_name,
-        &canonical_path,
+        package_path,
         package_source,
+        project_paths,
         provided,
         parents,
     )
@@ -700,6 +730,7 @@ fn provide_local_package(
 fn provide_git_package(
     _package_name: SmolStr,
     _repo: &str,
+    _project_paths: &ProjectPaths,
     _provided: &mut HashMap<SmolStr, ProvidedPackage>,
 ) -> Result<hexpm::version::Range> {
     let _git = ProvidedPackageSource::Git {
@@ -714,6 +745,7 @@ fn provide_package(
     package_name: SmolStr,
     package_path: &Path,
     package_source: ProvidedPackageSource,
+    project_paths: &ProjectPaths,
     provided: &mut HashMap<SmolStr, ProvidedPackage>,
     parents: &mut Vec<SmolStr>,
 ) -> Result<hexpm::version::Range> {
@@ -770,9 +802,17 @@ fn provide_package(
                     package_path.join(path)
                 };
                 // Recursively walk local packages
-                provide_local_package(name.clone(), &resolved_path, provided, parents)?
+                provide_local_package(
+                    name.clone(),
+                    &resolved_path,
+                    project_paths,
+                    provided,
+                    parents,
+                )?
             }
-            Requirement::Git { git } => provide_git_package(name.clone(), &git, provided)?,
+            Requirement::Git { git } => {
+                provide_git_package(name.clone(), &git, project_paths, provided)?
+            }
         };
         let _ = requirements.insert(name, version);
     }
@@ -794,9 +834,11 @@ fn provide_package(
 #[test]
 fn provide_wrong_package() {
     let mut provided = HashMap::new();
+    let project_paths = crate::project_paths_at_current_directory();
     let result = provide_local_package(
         "wrong_name".into(),
         &Path::new("./test/hello_world"),
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "subpackage".into()],
     );
@@ -814,10 +856,12 @@ fn provide_wrong_package() {
 #[test]
 fn provide_existing_package() {
     let mut provided = HashMap::new();
+    let project_paths = crate::project_paths_at_current_directory();
 
     let result = provide_local_package(
         "hello_world".into(),
         &Path::new("./test/hello_world"),
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "subpackage".into()],
     );
@@ -829,6 +873,7 @@ fn provide_existing_package() {
     let result = provide_local_package(
         "hello_world".into(),
         &Path::new("./test/hello_world"),
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "subpackage".into()],
     );
@@ -841,10 +886,11 @@ fn provide_existing_package() {
 #[test]
 fn provide_conflicting_package() {
     let mut provided = HashMap::new();
-
+    let project_paths = crate::project_paths_at_current_directory();
     let result = provide_local_package(
         "hello_world".into(),
         &Path::new("./test/hello_world"),
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "subpackage".into()],
     );
@@ -859,6 +905,7 @@ fn provide_conflicting_package() {
         ProvidedPackageSource::Local {
             path: Path::new("./test/other").to_path_buf(),
         },
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "subpackage".into()],
     );
@@ -872,9 +919,11 @@ fn provide_conflicting_package() {
 #[test]
 fn provided_is_absolute() {
     let mut provided = HashMap::new();
+    let project_paths = crate::project_paths_at_current_directory();
     let result = provide_local_package(
         "hello_world".into(),
         &Path::new("./test/hello_world"),
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "subpackage".into()],
     );
@@ -884,7 +933,7 @@ fn provided_is_absolute() {
     );
     let package = provided.get("hello_world").unwrap().clone();
     if let ProvidedPackageSource::Local { path } = package.source {
-        assert!(path.is_absolute())
+        assert!(!path.is_absolute())
     } else {
         panic!("Provide_local_package provided a package that is not local!")
     }
@@ -893,9 +942,11 @@ fn provided_is_absolute() {
 #[test]
 fn provided_recursive() {
     let mut provided = HashMap::new();
+    let project_paths = crate::project_paths_at_current_directory();
     let result = provide_local_package(
         "hello_world".into(),
         &Path::new("./test/hello_world"),
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "hello_world".into(), "subpackage".into()],
     );