+++ compiler-cli/src/main.rs
@@ -67,6 +67,7 @@ mod lsp;
 mod new;
 mod panic;
 mod publish;
+mod remove;
 mod run;
 mod shell;
 
@@ -205,6 +206,13 @@ enum Command {
         /// Add the packages as dev-only dependencies
         #[clap(long)]
         dev: bool,
+    },
+
+    /// Remove project dependencies
+    Remove {
+        /// The names of packages to remove
+        #[clap(required = true)]
+        packages: Vec<String>,
     },
 
     /// Clean build artifacts
@@ -423,6 +431,8 @@ fn main() {
         }
 
         Command::Add { packages, dev } => add::command(packages, dev),
+
+        Command::Remove { packages } => remove::command(packages),
 
         Command::Update => dependencies::update(),